+++ src/features/form/components/QuestionRenderers/FileUploadRenderer.tsx
@@ -1,47 +1,177 @@
 import * as formApi from "@/features/form/services/api";
 import type { FormsQuestionResponse } from "@nycu-sdc/core-system-sdk";
 import { RefreshCw, Upload, X } from "lucide-react";
-import { useRef, useState } from "react";
-import styles from "./FileUploadRenderer.module.css";
+import { useEffect, useRef, useState } from "react";
+import styles from "./FileUploadQuestion.module.css";
 
 interface FileItem {
 	id: string;
-	file: File;
-	status: "uploading" | "success" | "error";
+	file?: File;
+	filename: string;
+	serverInfo?: ServerFileInfo;
+	status: "loading" | "uploading" | "success" | "error";
 	error?: string;
 }
 
-interface FileUploadRendererProps {
+export interface ServerFileInfo {
+	fileId: string;
+	originalFilename: string;
+	contentType: string;
+}
+
+interface FileUploadQuestionProps {
 	responseId?: string;
 	questionId: string;
 	question: FormsQuestionResponse;
 	maxFileAmount: number;
 	allowedFileTypes: string;
 	onFilesChange: (fileNames: string) => void;
+	onFileMetadataChange?: (files: ServerFileInfo[]) => void;
 	disabled?: boolean;
+	initialFiles?: ServerFileInfo[];
 }
 
-export const FileUploadRenderer = ({ responseId, questionId, question, maxFileAmount, allowedFileTypes, onFilesChange, disabled = false }: FileUploadRendererProps) => {
+export const FileUploadQuestion = ({
+	responseId,
+	question,
+	questionId,
+	maxFileAmount,
+	allowedFileTypes,
+	onFilesChange,
+	onFileMetadataChange,
+	disabled = false,
+	initialFiles
+}: FileUploadQuestionProps) => {
 	const [items, setItems] = useState<FileItem[]>([]);
+	const itemsRef = useRef<FileItem[]>([]);
 	const inputRef = useRef<HTMLInputElement>(null);
+	const initialized = useRef(false);
 
-	const doUpload = async (item: FileItem) => {
-		if (!responseId) return;
-		setItems(prev => prev.map(i => (i.id === item.id ? { ...i, status: "uploading" as const, error: undefined } : i)));
+	useEffect(() => {
+		itemsRef.current = items;
+	}, [items]);
+
+	useEffect(() => {
+		if (initialized.current) return;
+		if (initialFiles === undefined) return; // data not yet loaded, wait
+		initialized.current = true;
+		if (initialFiles.length === 0) return;
+
+		const placeholders: FileItem[] = initialFiles.map(f => ({
+			id: crypto.randomUUID(),
+			filename: f.originalFilename,
+			serverInfo: f,
+			status: "loading" as const
+		}));
+		setItems(placeholders);
+
+		placeholders.forEach(async (placeholder, index) => {
+			const info = initialFiles[index];
+			try {
+				const res = await fetch(`/api/files/${info.fileId}`, { credentials: "include" });
+				if (!res.ok) throw new Error(`HTTP ${res.status}`);
+				const blob = await res.blob();
+				const file = new File([blob], info.originalFilename, { type: info.contentType || blob.type });
+				setItems(prev => {
+					const next = prev.map(i => (i.id === placeholder.id ? { ...i, file, status: "success" as const } : i));
+					onFilesChange(
+						next
+							.filter(i => i.status === "success")
+							.map(i => i.filename)
+							.join(",")
+					);
+					return next;
+				});
+			} catch {
+				setItems(prev => prev.map(i => (i.id === placeholder.id ? { ...i, status: "error" as const, error: "下載失敗，請重新上傳" } : i)));
+			}
+		});
+	}, [initialFiles]);
+
+	const doUploadBatch = async (batch: FileItem[]) => {
+		if (!responseId || batch.length === 0) return;
+		const batchIds = new Set(batch.map(i => i.id));
+		const existingFiles = itemsRef.current.filter(i => !batchIds.has(i.id) && i.file).map(i => i.file as File);
+		const newFiles = batch.map(i => i.file).filter((f): f is File => f !== undefined);
+		const allFiles = [...existingFiles, ...newFiles];
+		if (allFiles.length === 0) return;
+
+		setItems(prev => prev.map(i => (batchIds.has(i.id) ? { ...i, status: "uploading" as const, error: undefined } : i)));
 		try {
-			await formApi.uploadQuestionFiles(responseId, questionId, [item.file]);
+			const uploadResult = await formApi.uploadQuestionFiles(responseId, questionId, allFiles);
+			const serverInfoMap = new Map(uploadResult.files.map(f => [f.originalFilename, { fileId: f.fileId, originalFilename: f.originalFilename, contentType: f.contentType } as ServerFileInfo]));
 			setItems(prev => {
-				const next = prev.map(i => (i.id === item.id ? { ...i, status: "success" as const } : i));
+				const next = prev.map(i => ({
+					...i,
+					...(batchIds.has(i.id) ? { status: "success" as const } : {}),
+					serverInfo: serverInfoMap.get(i.filename) ?? i.serverInfo
+				}));
 				onFilesChange(
 					next
 						.filter(i => i.status === "success")
-						.map(i => i.file.name)
+						.map(i => i.filename)
 						.join(",")
 				);
+				const updatedServerInfos = next.filter(i => i.status === "success" && i.serverInfo).map(i => i.serverInfo as ServerFileInfo);
+				onFileMetadataChange?.(updatedServerInfos);
 				return next;
 			});
 		} catch (err) {
-			setItems(prev => prev.map(i => (i.id === item.id ? { ...i, status: "error" as const, error: (err as Error).message } : i)));
+			setItems(prev => prev.map(i => (batchIds.has(i.id) ? { ...i, status: "error" as const, error: (err as Error).message } : i)));
+		}
+	};
+
+	const doUpload = async (item: FileItem) => {
+		// If we already have a local File, just delegate to the batch uploader.
+		if (item.file) {
+			await doUploadBatch([item]);
+			return;
+		}
+
+		// For initially-loaded items where the download failed, try re-downloading
+		// the file from the server before uploading.
+		if (!item.serverInfo) {
+			// Nothing we can do without either a local file or server metadata.
+			return;
+		}
+
+		// Mark the item as loading while we re-fetch the file content.
+		setItems(prev => prev.map(i => (i.id === item.id ? { ...i, status: "loading" as const, error: undefined } : i)));
+
+		try {
+			const { fileId, originalFilename, contentType } = item.serverInfo;
+			const response = await fetch(`/api/files/${fileId}`);
+			if (!response.ok) {
+				throw new Error("下載失敗，請重新上傳");
+			}
+
+			const blob = await response.blob();
+			const downloadedFile = new File([blob], originalFilename, { type: contentType });
+
+			// Update the item in state with the newly created File.
+			let updatedItem: FileItem | null = null;
+			setItems(prev =>
+				prev.map(i => {
+					if (i.id !== item.id) return i;
+					updatedItem = {
+						...i,
+						file: downloadedFile,
+						filename: originalFilename
+					};
+					return updatedItem;
+				})
+			);
+
+			// If for some reason the item was not found in state, abort gracefully.
+			if (!updatedItem) {
+				throw new Error("找不到要重新上傳的檔案項目");
+			}
+
+			// Now that we have a local File, reuse the batch upload logic.
+			await doUploadBatch([updatedItem]);
+		} catch (err) {
+			const message = (err instanceof Error && err.message) || "下載失敗，請重新上傳";
+			setItems(prev => prev.map(i => (i.id === item.id ? { ...i, status: "error" as const, error: message } : i)));
 		}
 	};
 
@@ -53,35 +183,68 @@ export const FileUploadRenderer = ({ responseId, questionId, question, maxFileAm
 		const toAdd: FileItem[] = newFiles.slice(0, canAdd).map(file => ({
 			id: crypto.randomUUID(),
 			file,
+			filename: file.name,
 			status: "uploading" as const
 		}));
 		if (inputRef.current) inputRef.current.value = "";
 		setItems(prev => [...prev, ...toAdd]);
-		toAdd.forEach(doUpload);
+		doUploadBatch(toAdd);
 	};
 
-	const removeItem = (id: string) => {
-		setItems(prev => {
-			const next = prev.filter(i => i.id !== id);
+	const removeItem = async (id: string) => {
+		const originalItems = itemsRef.current;
+		const itemToRemove = originalItems.find(i => i.id === id);
+		const remaining = originalItems.filter(i => i.id !== id);
+		setItems(remaining);
+		onFilesChange(
+			remaining
+				.filter(i => i.status === "success")
+				.map(i => i.filename)
+				.join(",")
+		);
+		const remainingServerInfos = remaining.filter(i => i.status === "success" && i.serverInfo).map(i => i.serverInfo as ServerFileInfo);
+		onFileMetadataChange?.(remainingServerInfos);
+
+		if (!responseId || itemToRemove?.status !== "success") return;
+
+		try {
+			const remainingFiles = remaining.filter(i => i.status === "success" && i.file).map(i => i.file as File);
+			if (remainingFiles.length > 0) {
+				const uploadResult = await formApi.uploadQuestionFiles(responseId, questionId, remainingFiles);
+				const serverInfoMap = new Map(uploadResult.files.map(f => [f.originalFilename, { fileId: f.fileId, originalFilename: f.originalFilename, contentType: f.contentType } as ServerFileInfo]));
+				setItems(prev => {
+					const next = prev.map(i => ({ ...i, serverInfo: serverInfoMap.get(i.filename) ?? i.serverInfo }));
+					onFileMetadataChange?.(next.filter(i => i.status === "success" && i.serverInfo).map(i => i.serverInfo as ServerFileInfo));
+					return next;
+				});
+			} else {
+				await formApi.clearQuestionFiles(responseId, questionId);
+			}
+		} catch {
+			setItems(originalItems);
 			onFilesChange(
-				next
+				originalItems
 					.filter(i => i.status === "success")
-					.map(i => i.file.name)
+					.map(i => i.filename)
 					.join(",")
 			);
-			return next;
-		});
+			onFileMetadataChange?.(originalItems.filter(i => i.status === "success" && i.serverInfo).map(i => i.serverInfo as ServerFileInfo));
+		}
 	};
 
 	const activeCount = items.filter(i => i.status !== "error").length;
-	const canAddMore = !disabled && Boolean(responseId) && activeCount < maxFileAmount;
+	const isLoadingFromServer = items.some(i => i.status === "loading");
+	const canAddMore = !disabled && Boolean(responseId) && activeCount < maxFileAmount && !isLoadingFromServer;
 
 	return (
 		<>
 			<div className={styles.fileUploadList}>
 				{items.map(item => (
-					<div key={item.id} className={`${styles.fileItem} ${item.status === "error" ? styles.fileItemError : item.status === "uploading" ? styles.fileItemUploading : ""}`}>
-						<span className={styles.fileItemName}>{item.file.name}</span>
+					<div
+						key={item.id}
+						className={`${styles.fileItem} ${item.status === "error" ? styles.fileItemError : item.status === "uploading" || item.status === "loading" ? styles.fileItemUploading : ""}`}
+					>
+						<span className={styles.fileItemName}>{item.filename}</span>
 						{item.status === "uploading" && <span className={styles.fileItemHint}>上傳中…</span>}
 						{item.status === "error" && <span className={styles.fileItemHint}>{item.error || "上傳失敗"}</span>}
 						<div className={styles.fileItemActions}>
